[core][build] A collection of small detailed fixes and cleanups

+++ test/test_bonding.cpp
@@ -534,15 +534,15 @@ TEST(Bonding, Options)
 
     EXPECT_NE(srt_getsockflag(grp, SRTO_KMSTATE, &kms, &optsize), SRT_ERROR);
     EXPECT_EQ(optsize, (int) sizeof kms);
-    EXPECT_EQ(kms, int(SRT_KM_S_SECURED));
+    EXPECT_EQ(kms, int32_t(SRT_KM_S_SECURED));
 
     EXPECT_NE(srt_getsockflag(grp, SRTO_PBKEYLEN, &kms, &optsize), SRT_ERROR);
     EXPECT_EQ(optsize, (int) sizeof kms);
-    EXPECT_EQ(kms, 16);
+    EXPECT_EQ(kms, int32_t(16));
 
 #ifdef ENABLE_AEAD_API_PREVIEW
     EXPECT_NE(srt_getsockflag(grp, SRTO_CRYPTOMODE, &kms, &optsize), SRT_ERROR);
-    EXPECT_EQ(optsize, sizeof kms);
+    EXPECT_EQ(optsize, int(sizeof kms));
     EXPECT_EQ(kms, 1);
 #endif
 #endif
